refactor: migrate `skill` api as a server module.
