+++ internal/server/module/skill/handler.go
@@ -1,19 +1,28 @@
-package server
+package skill
 
 import (
 	"net/http"
+	"strconv"
 
 	"github.com/gin-gonic/gin"
 	"github.com/tingly-dev/tingly-box/internal/typ"
 )
 
-// ============================================
-// Skill Management API Handlers
-// ============================================
+// Handler handles skill-related HTTP requests
+type Handler struct {
+	manager *SkillManager
+}
+
+// NewHandler creates a new skill handler
+func NewHandler(manager *SkillManager) *Handler {
+	return &Handler{
+		manager: manager,
+	}
+}
 
 // GetSkillLocations returns all skill locations
-func (s *Server) GetSkillLocations(c *gin.Context) {
-	if s.skillManager == nil {
+func (h *Handler) GetSkillLocations(c *gin.Context) {
+	if h.manager == nil {
 		c.JSON(http.StatusInternalServerError, gin.H{
 			"success": false,
 			"error":   "Skill manager not initialized",
@@ -21,7 +30,7 @@ func (s *Server) GetSkillLocations(c *gin.Context) {
 		return
 	}
 
-	locations := s.skillManager.ListLocations()
+	locations := h.manager.ListLocations()
 
 	c.JSON(http.StatusOK, gin.H{
 		"success": true,
@@ -30,8 +39,8 @@ func (s *Server) GetSkillLocations(c *gin.Context) {
 }
 
 // AddSkillLocation adds a new skill location
-func (s *Server) AddSkillLocation(c *gin.Context) {
-	if s.skillManager == nil {
+func (h *Handler) AddSkillLocation(c *gin.Context) {
+	if h.manager == nil {
 		c.JSON(http.StatusInternalServerError, gin.H{
 			"success": false,
 			"error":   "Skill manager not initialized",
@@ -53,7 +62,7 @@ func (s *Server) AddSkillLocation(c *gin.Context) {
 		return
 	}
 
-	location, err := s.skillManager.AddLocation(req.Name, req.Path, req.IDESource)
+	location, err := h.manager.AddLocation(req.Name, req.Path, req.IDESource)
 	if err != nil {
 		c.JSON(http.StatusInternalServerError, gin.H{
 			"success": false,
@@ -70,8 +79,8 @@ func (s *Server) AddSkillLocation(c *gin.Context) {
 }
 
 // RemoveSkillLocation removes a skill location
-func (s *Server) RemoveSkillLocation(c *gin.Context) {
-	if s.skillManager == nil {
+func (h *Handler) RemoveSkillLocation(c *gin.Context) {
+	if h.manager == nil {
 		c.JSON(http.StatusInternalServerError, gin.H{
 			"success": false,
 			"error":   "Skill manager not initialized",
@@ -88,7 +97,7 @@ func (s *Server) RemoveSkillLocation(c *gin.Context) {
 		return
 	}
 
-	if err := s.skillManager.RemoveLocation(id); err != nil {
+	if err := h.manager.RemoveLocation(id); err != nil {
 		c.JSON(http.StatusInternalServerError, gin.H{
 			"success": false,
 			"error":   err.Error(),
@@ -103,8 +112,8 @@ func (s *Server) RemoveSkillLocation(c *gin.Context) {
 }
 
 // GetSkillLocation retrieves a specific skill location
-func (s *Server) GetSkillLocation(c *gin.Context) {
-	if s.skillManager == nil {
+func (h *Handler) GetSkillLocation(c *gin.Context) {
+	if h.manager == nil {
 		c.JSON(http.StatusInternalServerError, gin.H{
 			"success": false,
 			"error":   "Skill manager not initialized",
@@ -121,7 +130,7 @@ func (s *Server) GetSkillLocation(c *gin.Context) {
 		return
 	}
 
-	location, err := s.skillManager.GetLocation(id)
+	location, err := h.manager.GetLocation(id)
 	if err != nil {
 		c.JSON(http.StatusNotFound, gin.H{
 			"success": false,
@@ -137,8 +146,8 @@ func (s *Server) GetSkillLocation(c *gin.Context) {
 }
 
 // RefreshSkillLocation scans a location for updated skill list
-func (s *Server) RefreshSkillLocation(c *gin.Context) {
-	if s.skillManager == nil {
+func (h *Handler) RefreshSkillLocation(c *gin.Context) {
+	if h.manager == nil {
 		c.JSON(http.StatusInternalServerError, gin.H{
 			"success": false,
 			"error":   "Skill manager not initialized",
@@ -155,7 +164,7 @@ func (s *Server) RefreshSkillLocation(c *gin.Context) {
 		return
 	}
 
-	result, err := s.skillManager.ScanLocation(id)
+	result, err := h.manager.ScanLocation(id)
 	if err != nil {
 		c.JSON(http.StatusInternalServerError, gin.H{
 			"success": false,
@@ -172,8 +181,8 @@ func (s *Server) RefreshSkillLocation(c *gin.Context) {
 }
 
 // DiscoverIdes scans the home directory for installed IDEs with skills
-func (s *Server) DiscoverIdes(c *gin.Context) {
-	if s.skillManager == nil {
+func (h *Handler) DiscoverIdes(c *gin.Context) {
+	if h.manager == nil {
 		c.JSON(http.StatusInternalServerError, gin.H{
 			"success": false,
 			"error":   "Skill manager not initialized",
@@ -181,7 +190,7 @@ func (s *Server) DiscoverIdes(c *gin.Context) {
 		return
 	}
 
-	result, err := s.skillManager.DiscoverIdes()
+	result, err := h.manager.DiscoverIdes()
 	if err != nil {
 		c.JSON(http.StatusInternalServerError, gin.H{
 			"success": false,
@@ -197,8 +206,8 @@ func (s *Server) DiscoverIdes(c *gin.Context) {
 }
 
 // ImportSkillLocations imports discovered skill locations
-func (s *Server) ImportSkillLocations(c *gin.Context) {
-	if s.skillManager == nil {
+func (h *Handler) ImportSkillLocations(c *gin.Context) {
+	if h.manager == nil {
 		c.JSON(http.StatusInternalServerError, gin.H{
 			"success": false,
 			"error":   "Skill manager not initialized",
@@ -219,7 +228,7 @@ func (s *Server) ImportSkillLocations(c *gin.Context) {
 	}
 
 	imported := []typ.SkillLocation{}
-	existingLocations := s.skillManager.ListLocations()
+	existingLocations := h.manager.ListLocations()
 	existingPaths := make(map[string]bool)
 	for _, loc := range existingLocations {
 		existingPaths[loc.Path] = true
@@ -231,7 +240,7 @@ func (s *Server) ImportSkillLocations(c *gin.Context) {
 			continue
 		}
 
-		added, err := s.skillManager.AddLocation(loc.Name, loc.Path, loc.IDESource)
+		added, err := h.manager.AddLocation(loc.Name, loc.Path, loc.IDESource)
 		if err != nil {
 			// Log but continue with other locations
 			continue
@@ -242,12 +251,12 @@ func (s *Server) ImportSkillLocations(c *gin.Context) {
 	c.JSON(http.StatusOK, gin.H{
 		"success": true,
 		"data":    imported,
-		"message": "Imported " + string(rune(len(imported))) + " skill locations",
+		"message": "Imported " + strconv.Itoa(len(imported)) + " skill locations",
 	})
 }
 
 // GetSkillContent returns the content of a skill file
-func (s *Server) GetSkillContent(c *gin.Context) {
+func (h *Handler) GetSkillContent(c *gin.Context) {
 	locationID := c.Query("location_id")
 	skillID := c.Query("skill_id")
 	skillPath := c.Query("skill_path")
@@ -260,7 +269,7 @@ func (s *Server) GetSkillContent(c *gin.Context) {
 		return
 	}
 
-	if s.skillManager == nil {
+	if h.manager == nil {
 		c.JSON(http.StatusInternalServerError, gin.H{
 			"success": false,
 			"error":   "Skill manager not initialized",
@@ -268,7 +277,7 @@ func (s *Server) GetSkillContent(c *gin.Context) {
 		return
 	}
 
-	skill, err := s.skillManager.GetSkillContent(locationID, skillID, skillPath)
+	skill, err := h.manager.GetSkillContent(locationID, skillID, skillPath)
 	if err != nil {
 		c.JSON(http.StatusInternalServerError, gin.H{
 			"success": false,
@@ -285,8 +294,8 @@ func (s *Server) GetSkillContent(c *gin.Context) {
 
 // ScanIdes scans all IDE locations and returns discovered skills
 // This is a comprehensive scan that checks all default IDE locations
-func (s *Server) ScanIdes(c *gin.Context) {
-	if s.skillManager == nil {
+func (h *Handler) ScanIdes(c *gin.Context) {
+	if h.manager == nil {
 		c.JSON(http.StatusInternalServerError, gin.H{
 			"success": false,
 			"error":   "Skill manager not initialized",
@@ -294,7 +303,7 @@ func (s *Server) ScanIdes(c *gin.Context) {
 		return
 	}
 
-	result, err := s.skillManager.ScanIdes()
+	result, err := h.manager.ScanIdes()
 	if err != nil {
 		c.JSON(http.StatusInternalServerError, gin.H{
 			"success": false,
