+++ internal/server/module/skill/manager.go
@@ -1,4 +1,4 @@
-package data
+package skill
 
 import (
 	"crypto/sha256"
@@ -405,20 +405,21 @@ func resolvePath(homeDir, path string) string {
 // 2. First non-empty paragraph - returns the paragraph text (truncated if too long)
 //
 // Examples:
-//   Input: "# Commit Skill\n\nThis is a commit skill."
-//   Output: "Commit Skill"
 //
-//   Input: "## Code Review\n\nThis skill helps with reviews."
-//   Output: "Code Review"
+//	Input: "# Commit Skill\n\nThis is a commit skill."
+//	Output: "Commit Skill"
 //
-//   Input: "No heading here, just a description."
-//   Output: "No heading here, just a description."
+//	Input: "## Code Review\n\nThis skill helps with reviews."
+//	Output: "Code Review"
 //
-//   Input: "```go\nfunc main() {}\n```\n\nActual description here."
-//   Output: "Actual description here."
+//	Input: "No heading here, just a description."
+//	Output: "No heading here, just a description."
 //
-//   Input: "#   Spaced Heading  \n\nContent here."
-//   Output: "Spaced Heading"
+//	Input: "```go\nfunc main() {}\n```\n\nActual description here."
+//	Output: "Actual description here."
+//
+//	Input: "#   Spaced Heading  \n\nContent here."
+//	Output: "Spaced Heading"
 func parseSkillDescription(content string) string {
 	// Try to find first heading (# Heading)
 	headingRegex := regexp.MustCompile(`(?m)^#+\s*(.+)$`)
