+++ internal/server/module/skill/skill_client.json
